BaseController is implemented. Common navigation logic has been written in the basecontroller.js file.
