+++ webapp/controller/BaseController.js
@@ -1,0 +1,33 @@
+sap.ui.define([
+	"sap/ui/core/mvc/Controller",
+	"sap/ui/core/routing/History",
+	"sap/ui/core/UIComponent"
+], function (Controller, History, UIComponent) {
+	"use strict";
+
+	return Controller.extend("com.hardy.sapui5table.controller.BaseController", {
+
+		// ********************** Code Written By Naveen *******************
+		// ******* Navigate To Previous View *********
+		onPressBack: function (oEvent) {
+			var oHistory, sPreviousHash;
+			oHistory = History.getInstance();
+			sPreviousHash = oHistory.getPreviousHash();
+
+			if (sPreviousHash !== undefined) {
+				window.history.go(-1);
+			} else {
+				var oRouter = sap.ui.core.UIComponent.getRouterFor(this);
+				oRouter.navTo("homeview");
+			}
+
+		},
+
+		navToHome: function () {
+			var oRouter = sap.ui.core.UIComponent.getRouterFor(this);
+			oRouter.navTo("homeview");
+		}
+
+	});
+
+});
